requirements.txt
backports.pbkdf2==0.1
pycryptodome==3.12.0
requests==2.21.0
rich==11.0.0